jobsearchbe/requirements.txt
appdirs==1.4.4
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
distlib==0.3.0
Django==3.0.7
djangorestframework==3.11.0
filelock==3.0.12
idna==2.9
numpy==1.18.2
pytz==2020.1
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.8
virtualenv==20.0.21
